将 minidump 文件重命名为 crashdump, 因为 mini 前缀在 Linux 编译时有其他含义

+++ src/common/assistant.cpp
@@ -68,7 +68,7 @@ bool makeDirectories(const std::string& path) {
 	case ENOENT:
 		// parent didn't exist, try to create it
 	{
-		int pos = path.find_last_of('/');
+		size_t pos = path.find_last_of('/');
 		if (pos == std::string::npos)
 #if (defined(WIN32) || defined(_WIN32))
 			pos = path.find_last_of('\\');
